+++ src/auth-token-helpers/index.ts
@@ -1,4 +1,5 @@
 import { getTokenFromFile } from './get-token-from-file'
 import { getTokenUsingIam } from './get-token-using-iam'
+import { getTokenUsingUserPass } from './get-token-from-user-pass'
 
-export { getTokenFromFile, getTokenUsingIam }
+export { getTokenFromFile, getTokenUsingIam, getTokenUsingUserPass }