+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # UNRELEASED
 
+- Built using node.js 10
 - Add `getTokenUsingUserPass()` token helper
 - Fix a bug in the helpers if the token lease time exceeds a 32-bit signed int
 - Update `IBaseClient` object references to `Record<string, any>` type